ml-service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pdfplumber
python-docs
python-docx